Split persistent waveform cache entry schema
