+++ src/app/controller/playback/persistent_waveform_cache.rs
@@ -6,17 +6,18 @@
 //! session while still falling back safely when the source file changes.
 
 use super::*;
+mod entry;
+
 use crate::app::controller::playback::audio_cache::FileMetadata;
 use crate::app_dirs;
 use crate::sample_sources::SourceId;
-use crate::waveform::{DecodedWaveform, WaveformPeaks, next_cache_token};
-use serde::{Deserialize, Serialize};
+use crate::waveform::DecodedWaveform;
 use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-/// Increment when the serialized waveform cache schema changes incompatibly.
-const CACHE_VERSION: u32 = 1;
+use self::entry::{CACHE_VERSION, PersistentWaveformEntry};
+
 /// Extension used for persistent waveform cache payload files.
 const CACHE_FILE_EXTENSION: &str = "bin";
 
@@ -27,98 +28,6 @@ pub(crate) struct PersistentWaveformHit {
     pub(crate) decoded: Arc<DecodedWaveform>,
     /// Cached transient markers aligned with the decoded waveform payload.
     pub(crate) transients: Arc<[f32]>,
-}
-
-/// Serialized waveform cache entry stored on disk for one source path + metadata tuple.
-#[derive(Serialize, Deserialize)]
-struct PersistentWaveformEntry {
-    version: u32,
-    decoded: PersistentDecodedWaveform,
-    transients: Vec<f32>,
-}
-
-/// Serializable copy of `DecodedWaveform` without runtime-only cache identity.
-#[derive(Serialize, Deserialize)]
-struct PersistentDecodedWaveform {
-    samples: Vec<f32>,
-    analysis_samples: Vec<f32>,
-    analysis_sample_rate: u32,
-    analysis_stride: usize,
-    peaks: Option<PersistentWaveformPeaks>,
-    duration_seconds: f32,
-    sample_rate: u32,
-    channels: u16,
-}
-
-/// Serializable copy of waveform peak summaries used by the browser and waveform UI.
-#[derive(Serialize, Deserialize)]
-struct PersistentWaveformPeaks {
-    total_frames: usize,
-    channels: u16,
-    bucket_size_frames: usize,
-    mono: Vec<(f32, f32)>,
-    left: Option<Vec<(f32, f32)>>,
-    right: Option<Vec<(f32, f32)>>,
-}
-
-impl PersistentDecodedWaveform {
-    /// Clone one runtime decoded waveform into its persistent representation.
-    fn from_decoded(decoded: &DecodedWaveform) -> Self {
-        Self {
-            samples: decoded.samples.as_ref().to_vec(),
-            analysis_samples: decoded.analysis_samples.as_ref().to_vec(),
-            analysis_sample_rate: decoded.analysis_sample_rate,
-            analysis_stride: decoded.analysis_stride,
-            peaks: decoded
-                .peaks
-                .as_ref()
-                .map(|peaks| PersistentWaveformPeaks::from_peaks(peaks)),
-            duration_seconds: decoded.duration_seconds,
-            sample_rate: decoded.sample_rate,
-            channels: decoded.channels,
-        }
-    }
-
-    /// Restore one runtime decoded waveform with a fresh cache token.
-    fn into_decoded(self) -> DecodedWaveform {
-        DecodedWaveform {
-            cache_token: next_cache_token(),
-            samples: Arc::from(self.samples),
-            analysis_samples: Arc::from(self.analysis_samples),
-            analysis_sample_rate: self.analysis_sample_rate,
-            analysis_stride: self.analysis_stride,
-            peaks: self.peaks.map(|peaks| Arc::new(peaks.into_peaks())),
-            duration_seconds: self.duration_seconds,
-            sample_rate: self.sample_rate,
-            channels: self.channels,
-        }
-    }
-}
-
-impl PersistentWaveformPeaks {
-    /// Clone runtime waveform peaks into the persistent serialization shape.
-    fn from_peaks(peaks: &WaveformPeaks) -> Self {
-        Self {
-            total_frames: peaks.total_frames,
-            channels: peaks.channels,
-            bucket_size_frames: peaks.bucket_size_frames,
-            mono: peaks.mono.clone(),
-            left: peaks.left.clone(),
-            right: peaks.right.clone(),
-        }
-    }
-
-    /// Restore runtime waveform peaks from the serialized cache payload.
-    fn into_peaks(self) -> WaveformPeaks {
-        WaveformPeaks {
-            total_frames: self.total_frames,
-            channels: self.channels,
-            bucket_size_frames: self.bucket_size_frames,
-            mono: self.mono,
-            left: self.left,
-            right: self.right,
-        }
-    }
 }
 
 impl AppController {
@@ -185,14 +94,11 @@ pub(crate) fn load_persistent_waveform_cache_entry(
             return None;
         }
     };
-    if entry.version != CACHE_VERSION {
+    if entry.version() != CACHE_VERSION {
         let _ = std::fs::remove_file(&path);
         return None;
     }
-    Some(PersistentWaveformHit {
-        decoded: Arc::new(entry.decoded.into_decoded()),
-        transients: Arc::from(entry.transients),
-    })
+    Some(entry.into_hit())
 }
 
 /// Persist one waveform cache entry without requiring controller access.
@@ -213,11 +119,7 @@ pub(crate) fn persist_waveform_cache_entry(
             return;
         }
     };
-    let entry = PersistentWaveformEntry {
-        version: CACHE_VERSION,
-        decoded: PersistentDecodedWaveform::from_decoded(decoded),
-        transients: transients.as_ref().to_vec(),
-    };
+    let entry = PersistentWaveformEntry::from_runtime(decoded, transients);
     let bytes = match bincode::serialize(&entry) {
         Ok(bytes) => bytes,
         Err(err) => {
@@ -317,6 +219,7 @@ mod tests {
     use super::*;
     use crate::app::controller::test_support::dummy_controller;
     use crate::app_dirs::ConfigBaseGuard;
+    use crate::waveform::WaveformPeaks;
     use tempfile::tempdir;
 
     /// Build a compact decoded waveform fixture that exercises serialization fields.
